Modified the Makefile to allow for debug builds
The debug header is also modified to only print when in debug build.

Moved resources to main exe and moved config handling to main exe

Created a custom callback to handle hotkey presses.

Fixed a bug where I was not passing the correct value to RegisterHotkey

RegisterHotkey requires a virtual key scan code not a char as I initially believed.